# src/swagger.yaml
openapi: 3.0.0
info:
  title: WhatsApp Bridge API
  description: A RESTful API service that bridges WhatsApp Web with custom applications
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
    Status:
      type: object
      properties:
        authenticated:
          type: boolean
          description: Whether the client is authenticated
        initialized:
          type: boolean
          description: Whether the client is initialized
        qrDisplayed:
          type: boolean
          description: Whether a QR code is currently displayed
        state:
          type: string
          description: Current client state
        timestamp:
          type: integer
          description: Timestamp of the status check
    Account:
      type: object
      properties:
        me:
          type: string
          description: User's WhatsApp ID
        pushname:
          type: string
          description: User's display name
        wid:
          type: string
          description: User's WhatsApp ID (alternative format)
        platform:
          type: string
          description: Platform information
...
    MessageResponse:
      type: object
      properties:
        messageId:
          type: string
          description: ID of the sent message
        timestamp:
          type: integer
          description: Message timestamp
        status:
          type: string
          description: Message status (e.g., 'sent')
        from:
          type: string
          description: Sender's ID
        to:
          type: string
          description: Recipient's ID
    MessageDetails:
      type: object
      properties:
        id:
          type: string
          description: Message ID
        body:
          type: string
          description: Message content
        timestamp:
          type: integer
          description: Message timestamp
        from:
          type: string
          description: Sender's ID
        to:
          type: string
          description: Recipient's ID
        hasMedia:
          type: boolean
          description: Whether the message contains media
        type:
          type: string
          description: Message type
        isForwarded:
          type: boolean
          description: Whether the message is forwarded
        forwardingScore:
          type: integer
          description: Forwarding count
        isStatus:
          type: boolean
          description: Whether it's a status message
        isStarred:
          type: boolean
          description: Whether the message is starred
        broadcast:
          type: boolean
          description: Whether it's a broadcast message
        fromMe:
          type: boolean
          description: Whether the message is from the current user
        hasQuotedMsg:
          type: boolean
          description: Whether the message quotes another message
        mentionedIds:
          type: array
          items:
            type: string
          description: IDs of mentioned users
        links:
          type: array
          items:
            type: string
          description: URLs in the message
...
security:
  - BearerAuth: []
paths:
...
  /auth/status:
    get:
      summary: Get client status
      description: Get the current status of the WhatsApp client
      responses:
        '200':
          description: Current client status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/account:
    get:
      summary: Get connected account information
      description: Get information about the currently connected WhatsApp account
      responses:
        '200':
          description: Connected account information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          description: Client not fully initialized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /messages/reply:
    post:
      tags:
        - Messages
      summary: Reply to a message
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messageId
                - body
              properties:
                messageId:
                  type: string
                  description: ID of the message to reply to
                body:
                  type: string
                  description: Content of the reply message
      responses:
        200:
          description: Reply sent successfully
        400:
          description: Missing required fields
        404:
          description: Message not found
        500:
          description: Server error
  /messages/quote:
    post:
      tags:
        - Messages
      summary: Reply to a message with quote
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - messageId
                - body
              properties:
                messageId:
                  type: string
                  description: ID of the message to quote and reply to
                body:
                  type: string
                  description: Content of the reply message
      responses:
        200:
          description: Quote reply sent successfully
        400:
          description: Missing required fields
        404:
          description: Message not found
        500:
          description: Server error
  /messages:
    get:
      tags:
        - Messages
      summary: Get messages from a chat
      parameters:
        - name: chatId
          in: query
          required: true
          schema:
            type: string
          description: ID of the chat to get messages from
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
          description: Maximum number of messages to retrieve
      responses:
        200:
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MessageDetails'
        400:
          description: Missing chat ID
        500:
          description: Server error
  /messages/{messageId}/reply:
    post:
      tags:
        - Messages
      summary: Reply to a specific message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          description: ID of the message to reply to
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
                  description: Reply message content
                mediaUrl:
                  type: string
                  description: Optional URL for media attachment
      responses:
        200:
          description: Reply sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        400:
          description: Missing required fields
        404:
          description: Message not found
        500:
          description: Server error
  /messages/{messageId}/star:
    post:
      tags:
        - Messages
      summary: Star or unstar a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          description: ID of the message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                star:
                  type: boolean
                  default: true
                  description: Whether to star or unstar the message
      responses:
        200:
          description: Message starred/unstarred successfully
        404:
          description: Message not found
        500:
          description: Server error
  /messages/{messageId}/react:
    post:
      tags:
        - Messages
      summary: React to a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          description: ID of the message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - reaction
              properties:
                reaction:
                  type: string
                  description: Emoji reaction
      responses:
        200:
          description: Reaction added successfully
        400:
          description: Missing reaction emoji
        404:
          description: Message not found
        500:
          description: Server error
  /messages/{messageId}:
    delete:
      tags:
        - Messages
      summary: Delete a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          description: ID of the message to delete
        - name: everyone
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Whether to delete for everyone or just for me
      responses:
        200:
          description: Message deleted successfully
        404:
          description: Message not found
        500:
          description: Server error
  /messages/{messageId}/media:
    get:
      tags:
        - Messages
      summary: Download message media attachment
      description: Download media (image, video, audio, document) from a message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
          description: ID of the message containing media
      responses:
        200:
          description: Media file
          content:
            '*/*':
              schema:
                type: string
                format: binary
        400:
          description: Message has no media attachment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        404:
          description: Message or media not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'